Begins allowing different client types
