+++ src/server/clients/admin/admin-web-client.ts
@@ -1,0 +1,3 @@
+export class AdminWebClient {
+    playing: boolean = false;
+}
